♻️ 상품리스트 컴포넌트로 분리
메인페이지 및 카테고리 페이지의 공통로직 분리
