+++ pages/[category].tsx
@@ -1,13 +1,11 @@
-import Category from '../containers/Category';
 import { useRouter } from 'next/router';
+import ProductList from '../components/ProductList';
 
-const CategoryPage: React.FC = () => {
+export default function CategoryPage(): React.ReactElement {
   const router = useRouter();
   const { category } = router.query;
 
   const categoryString = typeof category === 'string' ? category : '';
 
-  return <Category category={categoryString} />;
-};
-
-export default CategoryPage;
+  return <ProductList category={categoryString} />;
+}
