+++ pages/[id].tsx
@@ -1,5 +1,5 @@
-import Detail from '../../containers/Detail';
 import { useRouter } from 'next/router';
+import Detail from '../components/Detail';
 
 const DetailPage: React.FC = () => {
   const router = useRouter();
